requirements.txt
discord.py
praw
python-dotenv
PyNaCl
spotipy
colorama
youtube_dl
youtube-search